+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 # Unreleased
 
-Known Issues:
-    * Run Scanner's interactive Swagger docs generating incorrect URLs in some configurations
+Changes:
+    * Fixed Run Scanner's interactive Swagger docs generating incorrect URLs in some configurations
 
 # 2.0.1